refactor(sync): extract writeback + storage/jobs/samples helpers (tech-debt)
Continues the sync_service.py split. No behavior change.

- `_sync_writeback.py`: manifest + local-file writeback after a push
  (writeback_create_config_in_manifest, writeback_create_row_in_manifest,
  propagate_kbc_metadata, writeback_after_push). Only writeback_after_push needs
  the service (for _write_config_file); the rest are pure.
- `_sync_storage.py`: pull-side data writers (write_storage_metadata,
  fetch_jobs_per_config, write_per_config_jobs, fetch_samples,
  mask_encrypted_columns) + the `_ensure_path_within` storage-write traversal
  guard moved here with its only callers. Only fetch_jobs_per_config needs the
  service (for _resolve_max_workers).

push()/pull() call the free functions; tests updated to the relocated symbols
(test_sync_service.py writeback calls, test_sync_storage_jobs.py storage calls).

`sync_service.py`: 3435 -> 2931 lines. Cumulative over this PR: 4033 -> 2931
(-1102, ~27%), now across 5 cohesive modules (_sync_models, _sync_bindings,
_sync_clone, _sync_writeback, _sync_storage). make check green (4092 passed,
identical to pre-split).
